Refactor enclave helper/bridge
- Rename avalon_enclave_bridge to avalon_enclave_info
- Move EnclaveHelper class from avalon_enclave_helper
  to avalon_enclave_info
- Rename class from EnclaveHelper to EnclaveInfo
- Update enclave_manager for refactor
