--- enclave_manager/avalon_enclave_manager/avalon_enclave_helper.py
@@ -1,111 +1,0 @@
-# Copyright 2018 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""
-This file defines the Enclave class to simplify integration of the Intel SGX
-enclave module into the rest of the Avalon flow. Typically, an application
-will call the initialize_enclave function first,
-then will call create_enclave_signup_data.
-"""
-
-import random
-import logging
-
-import avalon_enclave_manager.avalon_enclave_bridge as avalon_enclave
-import avalon_crypto_utils.keys as keys
-
-logger = logging.getLogger(__name__)
-
-
-# -----------------------------------------------------------------
-def initialize_enclave(enclave_config):
-    """initialize_enclave -- Call the initialization function on the
-    enclave module
-    """
-    return avalon_enclave.initialize_with_configuration(enclave_config)
-
-
-# -----------------------------------------------------------------
-class EnclaveHelper(object):
-    """
-    Wraps calls to the client for symmetry with the enclave service client
-    """
-
-    # -------------------------------------------------------
-    @classmethod
-    def create_enclave_signup_data(cls):
-        """
-        Creates enclave signup data
-        """
-
-        nonce = '{0:016X}'.format(random.getrandbits(64))
-        try:
-            enclave_data = \
-                avalon_enclave.create_signup_info(nonce)
-        except Exception as err:
-            raise Exception('failed to create enclave signup data; {}'
-                            .format(str(err)))
-
-        enclave_info = dict()
-        enclave_info['nonce'] = nonce
-        enclave_info['sealed_data'] = enclave_data.sealed_signup_data
-        enclave_info['verifying_key'] = enclave_data.verifying_key
-        enclave_info['encryption_key'] = enclave_data.encryption_key
-        enclave_info['encryption_key_signature'] = \
-            enclave_data.encryption_key_signature
-        enclave_info['enclave_id'] = enclave_data.verifying_key
-        enclave_info['proof_data'] = ''
-        if not avalon_enclave.enclave.is_sgx_simulator():
-            enclave_info['proof_data'] = enclave_data.proof_data
-
-        return cls(enclave_info)
-
-    # -------------------------------------------------------
-    def __init__(self, enclave_info):
-
-        # Initialize the keys that can be used later to
-        # register the enclave
-
-        try:
-            self.nonce = enclave_info['nonce']
-            self.sealed_data = enclave_info['sealed_data']
-            self.verifying_key = enclave_info['verifying_key']
-            self.encryption_key = enclave_info['encryption_key']
-            self.encryption_key_signature = \
-                enclave_info['encryption_key_signature']
-            self.proof_data = enclave_info['proof_data']
-            self.enclave_id = enclave_info['enclave_id']
-        except KeyError as ke:
-            raise Exception("missing enclave initialization parameter; {}"
-                            .format(str(ke)))
-
-        self.enclave_keys = \
-            keys.EnclaveKeys(self.verifying_key, self.encryption_key)
-
-    # -------------------------------------------------------
-    def send_to_sgx_worker(self, encrypted_request):
-        """
-        Submit workorder request to the Intel SGX Worker enclave
-
-        :param encrypted_request: base64 encoded encrypted workorder request
-        """
-        return avalon_enclave.send_to_sgx_worker(encrypted_request)
-
-    # -------------------------------------------------------
-    def get_enclave_public_info(self):
-        """
-        Return information about the enclave
-        """
-        return avalon_enclave.get_enclave_public_info(self.sealed_data)
-    # -------------------------------------------------------
